Represent positions with i32 rather than usize
Even though we can now represent the three empty negative quadrants,
it'll make the math simpler to implement so I figure there's no reason
not to do it.

+++ src/components.rs
@@ -6,8 +6,8 @@ use specs_derive::Component;
 /// Entities that have a physical position in the world.
 #[derive(Component)]
 pub struct Position {
-    pub x: usize,
-    pub y: usize,
+    pub x: i32,
+    pub y: i32,
 }
 
 /// Entities that need to be drawn as a single character.
